Move Star Wars API under starwars pages.
