+++ src/app/pages/pages.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 import { HomeComponent, AboutComponent, DraalAppHeaderComponent, DraalAppFooterComponent,
-    DraalAppRoutes, PlanetsComponent, SpeciesDetailComponent } from './index';
+    DraalAppRoutes, PlanetsComponent, SpeciesDetailComponent, StarWarsApiService } from './index';
 import { DraalDataTableModule, DraalAlertModule,
     DraalFormsModule, AlertComponent, DraalWidgetsCoreModule } from '../widgets';
 import { DraalServicesModule } from '../services';
@@ -22,6 +22,7 @@ import { DraalServicesModule } from '../services';
 
         DraalServicesModule.forRoot()
     ],
+    providers: [StarWarsApiService],
     declarations: [DraalAppHeaderComponent, DraalAppFooterComponent, HomeComponent,
         AboutComponent, PlanetsComponent, SpeciesDetailComponent],
     exports: [DraalAppHeaderComponent, DraalAppFooterComponent, AlertComponent]
