+++ src/app/pages/starwars/index.ts
@@ -1,2 +1,3 @@
 export * from './planets.component';
 export * from './species-detail.component';
+export * from './services';
