+++ src/app/services/services.module.ts
@@ -2,7 +2,7 @@ import { ModuleWithProviders, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { AppEventsService } from './events/appevent.service';
-import { ApiService, StarWarsApiService } from './api/api.service';
+import { ApiService } from './api/api.service';
 import { AlertService } from './alert/alert.service';
 import { NetworkService } from './network/network.service';
 import { AuthGuard } from './auth/auth.guard';
@@ -15,8 +15,7 @@ export class DraalServicesModule {
     static forRoot(): ModuleWithProviders {
         return {
             ngModule: DraalServicesModule,
-            providers: [AuthGuard, AppEventsService, ApiService, AlertService,
-                StarWarsApiService, NetworkService]
+            providers: [AuthGuard, AppEventsService, ApiService, AlertService, NetworkService]
         };
     }
 }
